+++ sbt-app/src/sbt-test/actions/doc-scala3/build.sbt
@@ -1,9 +1,2 @@
-lazy val rc1 = (project in file("rc1"))
-  .settings(
-    scalaVersion := "3.0.0-RC1"
-  )
-
+// Don't set scalaVersion
 lazy val a = project
-  .settings(
-    scalaVersion := "3.4.2",
-  )
