--- sbt-app/src/sbt-test/actions/doc-scala3/rc1/src/main/scala/foo/A.scala
@@ -1,10 +1,0 @@
-package foo
-
-final val foo="Foo"
-
-object A:
-  /**
-   * @param i An argument
-   */
-  def x(i:Int)=3
-
