+++ sbt-app/src/sbt-test/actions/doc-scala3/test
@@ -1,15 +1,7 @@
-> rc1 / doc
-
-# there shouldn't be two api/ directories
-# see https://github.com/lampepfl/dotty/issues/11412
-$ exists target/out/jvm/scala-3.0.0-RC1/rc1/api/api/index.html
-$ exists target/out/jvm/scala-3.0.0-RC1/rc1/api/api/foo/A$.html
-$ exists target/out/jvm/scala-3.0.0-RC1/rc1/api/api/foo.html
-
 > a / doc
 
 # there shouldn't be two api/ directories
 # see https://github.com/lampepfl/dotty/issues/11412
-$ exists target/out/jvm/scala-3.4.2/a/api/index.html
-$ exists target/out/jvm/scala-3.4.2/a/api/foo/A$.html
-$ exists target/out/jvm/scala-3.4.2/a/api/foo.html
+$ exists target/out/jvm/scala-3.5.2/a/api/index.html
+$ exists target/out/jvm/scala-3.5.2/a/api/foo/A$.html
+$ exists target/out/jvm/scala-3.5.2/a/api/foo.html
